error: remove type parameter to make it for general use
This is being done because by using type parameters ("generics")
the compiler will actually only compile `print_error` for
one of the structs that uses it, nowadays it is `ProcessError`,
but tomorrow it could be another one.

So to fix the issue I've used a reference to a trait, this
allows Rust to just know that it needs to point to someone
who has some methods, and that is safe, and more generalist.