io: make print_error function be generic
Before the function received a reference to a
trait, that is per say a trait object.

The problem of going to this approach instead
of using generics (typed parameters), is that
there is a cost at runtime instead of a cost
at size of the code compiled.

Since there is only one error type on the project
the drawback of generics is small, because Rust
will only compile `print_error` to the
`scoolite::error::Error` type.

Still, this optimization is minimal, the cost
was very little though.