command: remove IO side effects
Before the command module would print to the
stdout, however, by removing this and returning
Strings instead, makes the code pure and easier
to test.

We can call the functions of the module and
get the String result to see if what we wanted
happened. Instead of having to watch stdout.